backend/rag/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
qdrant-client==1.7.3
python-dotenv==1.0.0
transformers==4.35.2
torch==2.1.1
beautifulsoup4==4.12.2
PyYAML==6.0.1
markdown==3.5.1
google-generativeai==0.3.1
agents==0.1.0
openai==1.3.7